Convert `crystal.1` manpage to asciidoc (#15478)
This is a 1:1 translation of the manually crafted manpage to asciidoc. Also adds a simple make recipe to build the manpage.

The resulting man page is syntactically quite different, but the visuals are almost identical except for whitespace changes (some more empty lines for separation and less deep indentation).
I also fixed a couple of typographic errors while going through the document.

Previously we've done the same for shards: https://github.com/crystal-lang/shards/pull/262

This is a preparatory step for splitting into separate manpages (#15440).