Replace inline type IDs with global constants in LLVM IR (#15485)
Introduction and removal of types shifts the assignment of type IDs. They are numerical IDs in sequential order, with types defined later in the source code or the compiler receiving larger IDs than their sibling types.
Type IDs used to be inlined into the LLVM module, which can invalidate large portions of the object cache due to the change of type IDs.

This patch does not fight against the type ID assignment. It merely stops the inlining:

```llvm
@"Foo0:type_id" = external constant i32

; Function Attrs: uwtable
define ptr @"*Foo0@Reference::new:Foo0"() #0 {
; ...
  %1 = getelementptr inbounds %Foo0, ptr %0, i32 0, i32 0
  %2 = load i32, ptr @"Foo0:type_id", align 4
  store i32 %2, ptr %1, align 4
; ...
}
```

Global variables are required to be addressable in LLVM, so this creates an actual constant in the read-only section, hence the extra `load`. The actual compile-time value is now defined in `_main`:

```llvm
@"Foo0:type_id" = constant i32 6
```

With this simple trick, the object file cache can be reused more often.

This will hopefully improve build times in certain scenarios, such as rapid prototyping, and IDE integrations that run the whole compiler.