Do not load `Path` call receiver if known to be unused (#15488)
The compiler always emits LLVM instructions to load the receiver of a call even if it has a metaclass type, because there could be other side effects. This adds a large number of superfluous type ID loads when calling class methods:

```llvm
  %0 = load i32, ptr @"Foo0.class:type_id", align 4
  %1 = call ptr @"*Foo0@Reference::new:Foo0"()
  %2 = load i32, ptr @"Foo1.class:type_id", align 4
  %3 = call ptr @"*Foo1@Reference::new:Foo1"()
  %4 = load i32, ptr @"Foo2.class:type_id", align 4
  %5 = call ptr @"*Foo2@Reference::new:Foo2"()
  ; ...
```

Although `.new` does load the type ID later, this ID should never show up as an implicit argument to the `Foo*.new` methods. Here the `%0`, `%2`, and `%4` variables are a consequence of un-inlining the type IDs. This patch removes those loads when the call receiver is a `Path` known to refer to a non-generic metaclass type. (In isolation, the path skips over whatever is in `Crystal::CodeGenVisitor#visit(Crystal::Path)`.)