Fix handling of double dashes `--` in crystal `eval` command (#15477)
This allow to pass arguments after a double dash to the program when using the `eval` command.

```console
$ crystal eval 'puts ARGV' -- meow neigh
["meow", "neigh"]
```

Previously this caused a parser error (`Error: unexpected token: "meow"`) because the arguments after `--` were still considered code fragments.

This could be worked around by two double dashes, but it should really work with just a single one.

The reason for this behavior was that the `OptionParser` stops at `--` by default, and removes `--`. 
We need to use `unknown_args` to distinguish between arguments that were before `--` and arguments that were after `--`.

Co-authored-by: Johannes Müller <straightshoota@gmail.com>
Fixes: https://forum.crystal-lang.org/t/how-to-use-the-double-hyphen-in-the-eval-subcommand-of-crystal/7742