Fix parsing HTTP resource string that looks like absolute URL (#15499)
When parsing an HTTP resource string into a URI we expect this to be a relative URL, i.e. only the path, fragment and query components. Using `URI.parse` would misinterpret a path that looks like an absolute URL.

This patch introduces a new method `URI::Parser#parse_request_target`, mirroring `URI#request_target`. We use this to parse the request string of an HTTP request.
In the process I also simplified the URI parser a bit: We can pull up the recognition of a double-slash prefix directly to the entrypoint.

Co-authored-by: Julien Portalier <julien@portalier.com>