Must specific spawn context value (Fiber::ExecutionContext or nil)
The point is to avoid having fiber being spawn in an unknown context
from the isolated fiber. While we could default to `nil` it would delay
the understanding of the behavior to a runtime error, while having a
compilation error will bring up the issue sooner, and makes the intent
more explicit.

We'll decide later if we want to have a default value, and what (nil?
current context? default context?)