Isolated: lazy event loop initialization
An isolated context might not need an evloop and certainly doesn't need
one to be started upfront. It can be lazily instantiated on demand.

The fiber is isolated, so it doesn't need to bother with concurrency or
parallelism issues, however the context must support two mechanisms to
suspend the fiber/thread: wait on the event loop (when instantiated) and
wait on a condition variable otherwise.