Fix permissions application in `File.copy` (#15520)
Applying the final permissions already to a newly created file makes the extra call to `chmod` unnecessary.

We still need to detect whether the permissions are correct (this is the case when creating a new file or when an existing file already has the intended permissions). Otherwise we need to adjust the permissions. In that case this patch introduce an addional syscall (accessing `d.info` before `d.chmod`). There is no other way to determine whether `File.open` created a new file or opened an existing one.
This change ensures correct behaviour by avoiding unnecessary file mode changes.

The code comment claiming that we need to only change permissions after writing the data is based on a false assumption. Changing the permissions on disk does not have any effect on existing file descriptors. The OS only checks permissions when _opening_ a file.