Support `Slice.literal` in the interpreter (#15531)
This works by adding a specialized cache for the results of `pointerof` expressions that have the form `pointerof($Slice:n.@buffer)`, which can only arise from literal expansion of `Slice.literal` calls during the semantic phase.

As with string literals, the contents of slice literals are not truly read-only at the moment (#12422).