Avoids duplicate responses to ping frames
Before this change, when using `#on_ping` callback and defining your
own block of code to handle it, resulted in two `pong` frames being sent
to the other site of the WebSocket.

This change setups a default ping <-> pong response cycle but once you
override it, it will use yours instead.