Fix the return type of `Enumerable#sum/product` for union elements (#15314)
Using the first type of a union type as the type of the result of `Enumerable#sum/product()` call can cause runtime failures, e.g. `[1, 10000000000_u64].sum` causes `OverflowError`. A safer alternative is to disallow the use of union types with `Enumerable#sum`, `#produc` and recommend the use of `#sum(initial)`, `#product(initial)` with an initial value of the expected result type.

Co-authored-by: Venkatesh-Prasad Ranganath <rvprasad@users.noreply.github.com>
Co-authored-by: Johannes Müller <straightshoota@gmail.com>
Co-authored-by: Sijawusz Pur Rahnama <sija@sija.pl>