Add `Fiber::ExecutionContext::SingleThreaded` scheduler (#15511)
Introduces the first EC scheduler that runs in a single thread.

Uses the same queues (`Runnables`, `GlobalQueue`) as the multithreaded scheduler that will come next. The `Runnables` local queue could be simplified (no parallel accesses, hence no need for atomics) at the expense of duplicating the implementation.

The scheduler doesn't need to actively park the thread; the event loops always block (when told to) even when there are no events, which acts as parking the thread.

NOTE: we can start running the specs in this context (they pass locally), though there might be some issues until we replace rogue `Thread` instances with bubble wrapped `ExecutionContext::Isolated` (coming in another PR). If you want to try:

```console
$ make std_spec FLAGS="-Dpreview_mt -Dexecution_context"
$ make compiler_spec FLAGS="-Dpreview_mt -Dexecution_context"
```

Co-authored-by: Johannes Müller <straightshoota@gmail.com>