Support `sizeof` and `alignof` inside macros for stable types (#15497)
This patch enables the use of `sizeof` and `alignof` in macro expressions when the type's size and alignment are known to be __stable__, i.e. their values cannot be changed by adding new code. The idea was raised in [this forum thread](https://forum.crystal-lang.org/t/byte-size-of-a-type-in-a-macro-i-e-cant-call-sizeof-in-a-macro/6549) and possibly in other GitHub issues as well. With this PR you can write compile-time checks like:

```crystal
{% unless sizeof(Void*) == sizeof(LibC::SizeT) %}
  {% raise "we don't allow the x86 Real Mode here" %}
{% end %}
```

All types are __stable__, except for the following:

* Non-primitive structs, e.g. `Bytes`
* `ReferenceStorage` instances
* Modules, e.g. `GC` (not to be confused with `GC.class`, all metaclasses are stable)
* Uninstantiated generic types, e.g. `Array`
* Annotations, e.g. `x = uninitialized Flags` (these "types" are ill-behaved anyway)
* `StaticArray`, `Tuple`, `NamedTuple`, and `Union` instances containing any unstable constituent types

`typeof` is disallowed, because we obviously cannot do any kind of non-macro semantic analysis inside macros.

Although the non-macro `sizeof` and `alignof` support constants, their macro counterparts do not, even when the constant evaluates to a simple compile-time literal (`A = 1; {{ sizeof(A) }}`).