Add `Crystal::EventLoop#sleep(duration)` method (#15564)
The event loop interface for sleeping has been inherited from the libevent event loop. It needs to allocate an event, stores on Fiber to avoid repeated allocations, and we mustn't forget to free the event before the fiber terminates.

But none of the new event loops (iocp, epoll, kqueue or the incoming io_uring) need to allocate or free anything :confused: 

With the new `sleep(Time::Span)` method, event loops behaves just like the other event loop methods: it blocks the current fiber and returns when ready. This removes the indirect "create event object" then interact with that indirect object to have the event loop sleep. Now the event loop is fully responsible for sleeping.

**Note 1**: I kept `Fiber#resume_event` for compatibility reasons (i.e. free the libevent allocation when the fiber terminates). We might consider to only declare it when `Crystal::EventLoop::LibEvent` is defined or have a mechanism to declare fiber local values.

**Note 2**: `Fiber#timeout_event` is different. It handles the special case of `select` timeouts, where a fiber might be resumed _twice_ (by `channel.send` and `sleep`) and arming the timer mustn't suspend the current fiber.