Add `Fiber::ExecutionContext::Isolated` (#15513)
Aka bubble wrap a thread.

The `Fiber::ExecutionContext::Isolated` context is a special scheduler: it starts a single thread to execute a block of code inside its main fiber. Concurrency is disabled, no other fiber can run on this thread. The fiber owns the thread and can do any number of blocking operations as it needs to without blocking any other fiber.

But you can still:

- do IO operations normally;
- communicate with other fibers (in other contexts) normally through the regular sync primitives (`Channel`, `Mutex` or `WaitGroup`);
- `spawn` fibers: they will be spawned into another context (`ExecutionContext.default` by default).

Co-authored-by: Johannes Müller <straightshoota@gmail.com>