ARM: reduce duplication in fiber context switch (#15585)
Instead of duplicating the ASM twice (and twice again for LLVM < 9), we can put a few macros in the ASM string to conditionally generate the floating-point instructions.

Note: This only works because the ASM is within a wrapping macro, all the macros get resolved and only _then_ does the compiler notices the `asm` directive.