Optimize `Path#anchor` (#15584)
Factors out the body of `#drive_and_root` to return the end indices of the drive and root parts respectively, instead of constructing substrings immediately.

When both the drive and root are present, which happens to all Windows absolute or UNC paths, this saves a string allocation in `#drive` / `#root`, and a string concatenation in `#anchor`. This also emphasizes by construction the invariant that `#drive` and `#anchor` are always prefixes of the whole path if they exist.