Respect `--mcpu=help` in the compiler (#15595)
Passing `--mcpu=help` when building or running any Crystal source code makes LLVM display a list of all valid CPU and feature flags for the currently selected target, but then Crystal will continue with the semantic phase before failing to create a target machine for an invalid CPU. This PR makes the CLI immediately exit instead when that happens.

`--mattr=+help` can be used to display the same help message without exiting the compiler nor adding an invalid feature flag; nothing needs to be done in Crystal about this.