Add thread safety to `at_exit` (#15598)
Protects the `@@handlers` class variable of `Crystal::AtExitHandlers` with a thread mutex. 

Avoids a parallel initialization that could create two distinct arrays, and lose one `at_exit` handler;

Avoids a parallel mutation of the array when `at_exit` is called in parallel to running the handlers, ~~or when `at_exit` is called by an `at_exit` handler :shrug:~~ In these edge cases, the handler is merely added and will be run next.

There should be no behavior change from not using the mutex compared to ST.

I didn't branch based on the `preview_mt` flag because in single-threaded the mutex is usually not contended, and its impact is negligible overall. That keeps the implementation simple.