Fix error message when `PollDescriptor` can't transfer fd (#15663)
Removes the `BUG:` prefer in the RuntimeError message when we can't transfer a `fd` to another event loop.

The prefix usually denotes a broken expectation in the stdlib implementation, which means the stdlib did a mistake and we shall fix it, but this isn't the case here.

An example for a BUG is just a few lines above to catch trying to transfer to the _same_ event loop (it makes no sense).