Fix `libgc` pkg-config name for version discovery (#15636)
We define different names for `pkg-config` in the `Link` annotations of `LibGC` depending on target, but used a fixed name in version discovery.
This patch re-uses the name declared in the annotation. We need to filter for the appropriate annotation because there are multiple ones. Unfortunately, I haven't found an easier way to do this.

Co-authored-by: Quinton Miller <nicetas.c@gmail.com>