optimize rfc4007 parsing approach, unify exceptions
Instead of using `#partition('%')` on the input string, which results in
additional `String` allocations, we can stop consuming the slice within
`#parse_v6_fields?` when we encounter the zone seprator character (`'%'`)
and pass that index back to the constructor.

However, this change to `#parse_v6_fields?` is a bit more intrusive and
changes it return type from a simple `UInt16[8]?` to a
`Tuple(UInt16[8]?, Int32?)` to account for the zone separator index.

All references to that method and its specs were updated accordingly to
adopt this change.

Additionally the exception conditions for invalid v6 addresses have been
simplified a bit and all of the converted to proper `Socket::Error`
exceptions.