rework rfc4007 parsing to subslice
This is proposed as an alternate solution for the previous
implementation.
Returning the index of the zone separator (`'%'`) back to the
constructor is a bit cumbersome for subsequent usage.
Return the remaining subslice instead makes it much easier to use in the
constructor for subsequent parsing and within the exception messages.

I definitely prefer this approach over the previous one and consider it
an improvement.