Add `Crystal::EventLoop#reopened(FileDescriptor)` (#15640)
We don't want to _close_ the file descriptor after we reopened it. Instead, some event loops want to resume the pending operations on the old file descriptor, so they can be restarted on the new file descriptor and at worst wait on the new one.

`#reopened(FileDescriptor)` explicits this difference in intended behavior.

It doesn't change anything for the current event loops, but `io_uring` for example can close asynchronously so we must make the distinction, because we definitely don't want to close after reopen :sweat_smile: 

A follow-up will move the actual close to the eventloop and `#close(io)` will do what it's supposed to do (close the file descriptor).