Support Windows local device paths in `Path` (#15590)
There are [7 types of user-space paths on Win32](https://googleprojectzero.blogspot.com/2016/02/the-definitive-guide-on-win32-to-nt.html):

* Drive-absolute (`C:\foo`)
* Drive-relative (`C:foo`)
* Rooted (`\foo`)
* Relative (`foo`)
* UNC absolute (`\\server\share\foo`)
* Local device (`\\.\NUL`, `\\?\C:\foo`)
* Root local device (`\\.`, `\\?`)

Local device paths starting with exactly `\\?\` are also known as raw local device paths. Other than this prefix, all forward and backward slashes are interchangeable.

`Path` already recognizes the first 5 of them; this PR adds the remaining 2. As noted in #15587, the `\\.\` prefix appears when dealing with special devices, such as `\\.\NUL` and `\\.\COM15`, as well as named pipes. The `\\?\` prefix is useful when passing paths to Win32 functions that are already normalized, or if the path exceeds 260 UTF-16 code units (see also #13420).

This does not cover NT paths of the form `\??\...`. They appear in the reparse data area of Windows symbolic links.