Optimize `Indexable#find` (#15674)
The call to `index` performs the bounds check twice and calls `unsafe_fetch` for the found element again. It's better to iterate and return the element directly on a match.

A further enhancement could perhaps extract the shared implementation of `index` and `find` into `find_with_index` (which might or might not be a public method).