Make `offset` a required parameter in `Indexable#find` (#15671)
Since 1.16.0, `Indexable#find` provides the `offset` parameter as an additional feature over `Enumerable#find`.
Calls that do not make use of this parameter don't need the added complexity of offset handling. They should fall back to the base implementation in `Enumerable#find`.

To achieve that, this patch makes `offset` a required parameter for  `Indexable#find`.