Add `Time.month_week_date` (#15620)
This function returns a `Time` instance using a given month, week of month, and day of week. It is similar to the existing `.week_date`, except the week number is relative to the given month, not the beginning of the given year (or the end of the previous year).

Windows time zone transitions are specified in this manner, as are the [`M` dates of POSIX `TZ` strings](https://pubs.opengroup.org/onlinepubs/9799919799/basedefs/V1_chap08.html).