Fix: CRYSTAL_LOAD_DEBUG_INFO=1 fails with -Dexecution_context (#15704)
Loading the debug info relies on `File` and goes through the event loop (even if reading files is blocking) which in turn requires a fiber scheduler (event loops are tied to a scheduler).

They used to be lazily instantiated, but with execution contexts it must be explicit. We thus have to load the debug info after creating the default execution context.