Use `Slice.literal` for `fast_float` when supported (#15667)
A compiler version of 1.16.0 or above indicates support for slice literals, including in the interpreter, so we can start switching to them and dropping the old dynamic array initializers.

This is of course not the largest lookup table in the standard library, so the build time and binary size savings are rather limited. This table is chosen to give a taste of what's possible in subsequent patches.