Fix: raise on manual fiber resume from sleep
We can suspend a fiber by calling `sleep(time)` and we expect the fiber
to only be resumed when the sleep time expires, but there is nothing
preventing to enqueue the fiber (because of errors), which will
unexpectedly resume the fiber early.

This leads to at best errors (`wake_at` can't be nil) as well as
segfaults because the event, allocated on the stack, is still in timers
the sleep method returned, or to double resumes, ...

This should usually not happen, unless there is a programming error in
the runtime, but someone wants to use `sleep` as a timeout mechanism
(it's not).