Open file with blocking: nil or false in another thread
We can't reliably detect without significant overhead whether the file at
*path* might block for a while before calling open; while O_NONBLOCK has
no effect on regular disk files, special file types are a different story.

Open with O_NONBLOCK will fail with ENXIO for O_WRONLY (no connected
reader) but it will always succeed for O_RDONLY (regardless of a connected
writer or not), then any attempt to read will return EOF, leaving no means
to wait until a writer connects.

We thus rely on the *blocking* arg: when false the file might be a special
file type, so we check it; if it's a fifo (named pipe) or a character
device, we open in another thread so we don't risk blocking the current
thread (and thus other fibers) until a reader or writer is also connected.

We need preview_mt to safely re-enqueue the current fiber from the thread.