Do not use private linkage for slice literal buffers (#15746)
The following snippet used to fail because a method in `Foo`'s LLVM module refers to a `$Slice:*` buffer variable defined in the main LLVM module:

```crystal
module Foo
  def self.foo
    Slice.literal(1)
  end
end

Foo.foo
```

The correct linkage is now used for the `$Slice:*` variables.

The existing workaround is to move the literal to a constant, in which case the constant initializer (`const_init` or `const_read`) is also defined in the main LLVM module, thus eliminating the link-time failure. But see also #15701.