Fix raise on manual fiber resume from sleep (#15744)
We can suspend a fiber by calling `sleep(time)` and we expect the fiber to only be resumed when the sleep time expires, but there is nothing preventing to enqueue the fiber (because of errors), which will unexpectedly resume the fiber early.

This leads to at best errors (`wake_at` can't be nil) as well as segfaults because the event, allocated on the stack, is still in timers the sleep method returned, or to double resumes, ...

This should usually not happen, unless there is a programming error in the runtime, or someone wants to use `sleep` as a timeout mechanism (it's not).

I discovered this the hard way while implementing sync primitives. Having an explicit exception is much nicer than segfaults and timer related errors because `Fiber.yield` got manually resumed.