Support debug info of 128-bit enum members (#15770)
On the LLVM 21 development branch, the `LLVMDIBuilderCreateEnumeratorOfArbitraryPrecision` function (which I submitted to LLVM) supports enumerators larger than the 64 bits that `LLVMDIBuilderCreateEnumerator` can handle.

The underlying C++ implementation can be backported all the way back to LLVM 13, but if we added it to `llvm_ext.cc`, this would create an awkward situation where version 18 to 20 would miss out for no apparent reason. Thus, this PR requires LLVM 21 or above.