[CI] Add CI workflow for MinGW-w64 ARM64 builds (#15794)
This is mostly just a copy of the workflow definition before #15453, slightly adjusted for ARM64. The cross-compilation itself does not require an AArch64 host machine, and an x86-64 will just work equally fine. Once 1.17.0 is out and a release package is available, we could drop this again and use a job matrix instead.