Preserve precision when passing `BigDecimal` or `BigFloat` to `sprintf` `%i` (#15808)
When `String::Formatter` formats a non-integer for integer specifiers such as `%d`, it calls `#to_i` on that value to reuse the implementation for `Int32`, meaning `BigDecimal`s and `BigFloat`s outside `Int32`'s range will raise an `OverflowError`. Fortunately, `BigInt` is already compatible with `sprintf`, so they could delegate to `BigInt` instead of `Int32`.

This may look similar to #15203, but the fix there is not as simple as delegating to `#to_big_f`, because that also loses precision.