Add Crystal::EventLoop#pipe
Moves the creation of IO pipes to the event loop, so they decide to set
the underlying file descriptor or handle to blocking or non blocking
mode.

This is further reflected by the read_blocking and write_blocking args
that now default to nil, instead of false, so the event loop will use
its own requirements.

This is a delayed breaking change: the current event loops will all set
O_NONBLOCK on the file descriptors, or FILE_FLAG_OVERLAPPED for IOCP, by
default, but future event loops can break this assumption (io_uring).