Simplify `Crystal::System::Fiber::RESERVED_STACK_SIZE` initializer on Windows (#15820)
`LibC::DWORD.new(0x10000)` is a complex initializer, which on very rare
occasions could completely halt the program:

Maybe it was interfering with the event loop before fiber stacks could be
checked out? This was especially observable on the recent ARM64 MSYS2 CI, but if
you run a Hello World program in a shell loop the same also happens in a matter
of seconds.