Add Fiber::ExecutionContext::ThreadPool
A global pool of thread to start new threads from, and return threads
to, so we don't start and stop threads all the time, and can wake an
existing thread instead of creating a new one from scratch.

The thread pool still eventually shuts down a thread after a
configurable keepalive is reached, but takes extra measures to never
shutdown the main thread, which would invalide the program's main fiber
stack (segfaults).