IO::FileDescriptor: improve docs around raw fd/handle
Explicits that the fd/handle will be configured to match the event loop
requirements when adopted in the Socket types (its configuration will be
changed).

Explicits that `#fd` returns the fd/handle configured for the current
event loop, and no specific configuration must be expected.

Explicits that changing a socket's blocking mode can break the event
loop, and as such must be treated carefully.