Add Fiber::Execution.syscall(&)
Marks the scheduler has running a blocking syscall.

The monitor thread now ticks every 10ms to check if any scheduler in any
ST or MT context is blocked on a syscall, and if so tries to detach the
scheduler from the thread. On success the scheduler is moved to another
thread, taken from the thread pool.

The fiber doing a blocking syscall will still be blocked, but other
fibers may be resumed by the scheduler.

When the blocking syscall returns, the thread will try to unmark the
scheduler as running a blocking syscall. Upon success it returns. Upon
failure, it enqueues the current fiber back into its execution context,
and checks itself back into the thread pool.