Add `truncate_at_null` parameter to `String.new(Bytes)` and `.from_utf16` (#15887)
It is common to find C strings in `StaticArray`s or `Slice`s that are null-terminated, but may also take up the entire slice if its size is equal to the slice's size. In this case, `String.new(UInt8*)` is unsafe, whereas `String.new(Bytes)` followed by `rstrip('\0')` is not very efficient. This PR allows the `Bytes` constructor to stop as soon as it encounters a null character.

`LibC::SockaddrUn#sun_path` is not necessarily null-terminated in practice, according to https://man7.org/linux/man-pages/man7/unix.7.html. `LibC::Dirent#d_name` is always null-terminated (in fact, it is declared as a variable-length array on Solaris).

There are probably more places where this is applicable to Win32 code than the ones included here.