Add ameba (#15875)
Adds configuration and a CI job for https://github.com/crystal-ameba/ameba

I let ameba generate a config (`ameba --gen-config`) but replaced all exclusions with hard disables. Then I fixed the violations for some of the rules, which have already been merged in the last couple of days. So I removed the disables for these rules.

The configuration is annotated with explanations in code.

I have disabled most `Naming` and `Style` rules because most had large numbers of violations and some of them seem questionable. We can consider enabling some of the in the future, piece by piece