Fix interpreter guard clauses for signal handling (#15892)
The guard clauses introduced in #15178 to protect using the API for https://github.com/crystal-lang/crystal/issues/14766 are wrong. The guards check for instance methods, which are never defined because we're using class methods instead.
As a result, the fix for #12241 was never actually applied.
The spec file uses the same guard clause, so it didn't run.