Extract XML::Document from XML::Node
Introduces the XML::Document type to wrap a Pointer(LibXML::Doc), while
XLM::Node is meant to wrap the subtree Pointer(LibXML::Node). The
XML::Document inherits from XML::Node and acts exactly like a XML::Node,
so it should hopefully be a transparent change.

The main advantage is that we now register a single finalizer instead of
one for every node (only the document needed a finalizer). Sadly, it
doesn't allow to have a dual strong reference from a document to its
nodes, and from the nodes to their document. Boehm GC still complains
about cyclic finalization, despite only the document having a
finalizer.